+++ src/library/DIPS.Mobile.UI/Components/TextFields/InputFields/MultiLineInputField/MultiLineInputField.Properties.cs
@@ -32,6 +32,65 @@ public partial class MultiLineInputField
         get => (ICommand?)GetValue(SaveCommandProperty);
         set => SetValue(SaveCommandProperty, value);
     }
+
+    /// <summary>
+    /// The parameter to be sent with <see cref="SaveCommand"/>
+    /// </summary>
+    public object SaveCommandParameter
+    {
+        get => (object)GetValue(SaveCommandParameterProperty);
+        set => SetValue(SaveCommandParameterProperty, value);
+    }
+
+    public bool IsSavingSuccess
+    {
+        get => (bool)GetValue(IsSavingSuccessProperty);
+        set => SetValue(IsSavingSuccessProperty, value);
+    }
+    
+    /// <summary>
+    /// Whether the component is saving or not, will fade out text and display a spinner
+    /// </summary>
+    public bool IsSaving
+    {
+        get => (bool)GetValue(IsSavingProperty);
+        set => SetValue(IsSavingProperty, value);
+    }
+    
+    /// <summary>
+    /// Determines whether the component is in an error state or not
+    /// </summary>
+    public bool IsError
+    {
+        get => (bool)GetValue(IsErrorProperty);
+        set => SetValue(IsErrorProperty, value);
+    }
+
+    /// <summary>
+    /// Sets the error text that is displayed as HelpText
+    /// </summary>
+    public string ErrorText
+    {
+        get => (string)GetValue(ErrorTextProperty);
+        set => SetValue(ErrorTextProperty, value);
+    }
+    
+    public static readonly BindableProperty ErrorTextProperty = BindableProperty.Create(
+        nameof(ErrorText),
+        typeof(string),
+        typeof(MultiLineInputField));
+    
+    public static readonly BindableProperty IsErrorProperty = BindableProperty.Create(
+            nameof(IsError),
+            typeof(bool),
+            typeof(MultiLineInputField),
+            propertyChanged: (bindable, _, _) => ((MultiLineInputField)bindable).OnIsErrorChanged());
+    
+    public static readonly BindableProperty IsSavingProperty = BindableProperty.Create(
+        nameof(IsSaving),
+        typeof(bool),
+        typeof(MultiLineInputField),
+        propertyChanged: (bindable, _, _) => ((MultiLineInputField)bindable).OnIsSavingChanged());
     
     public static readonly BindableProperty SaveCommandProperty = BindableProperty.Create(
         nameof(SaveCommand),
@@ -49,4 +108,15 @@ public partial class MultiLineInputField
         typeof(bool),
         typeof(MultiLineInputField),
         defaultBindingMode:BindingMode.OneWayToSource);
+    
+    public static readonly BindableProperty IsSavingSuccessProperty = BindableProperty.Create(
+        nameof(IsSavingSuccess),
+        typeof(bool),
+        typeof(MultiLineInputField),
+        propertyChanged: (bindable, _, _) => ((MultiLineInputField)bindable).OnIsSavingSuccessChanged());
+    
+    public static readonly BindableProperty SaveCommandParameterProperty = BindableProperty.Create(
+        nameof(SaveCommandParameter),
+        typeof(object),
+        typeof(MultiLineInputField));
 }