+++ src/library/DIPS.Mobile.UI/Components/TextFields/InputFields/MultiLineInputField/MultiLineInputField.cs
@@ -1,8 +1,11 @@
+using DIPS.Mobile.UI.Components.CheckBoxes;
 using DIPS.Mobile.UI.Resources.LocalizedStrings.LocalizedStrings;
 using DIPS.Mobile.UI.Resources.Styles;
 using DIPS.Mobile.UI.Resources.Styles.Button;
+using DIPS.Mobile.UI.Resources.Styles.InputField;
 using DIPS.Mobile.UI.Resources.Styles.Label;
 using Button = DIPS.Mobile.UI.Components.Buttons.Button;
+using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 using HorizontalStackLayout = DIPS.Mobile.UI.Components.Lists.HorizontalStackLayout;
 
 namespace DIPS.Mobile.UI.Components.TextFields.InputFields.MultiLineInputField;
@@ -13,7 +16,7 @@ public partial class MultiLineInputField : SingleLineInputField
 
     private string? m_textWhenFirstFocused;
 
-    private readonly Label m_label = new Labels.Label() 
+    private readonly Label m_label = new Labels.Label
     { 
         Style = Styles.GetLabelStyle(LabelStyle.Body200), 
         IsVisible = false,
@@ -21,11 +24,25 @@ public partial class MultiLineInputField : SingleLineInputField
         VerticalTextAlignment = TextAlignment.Start,
         LineBreakMode = LineBreakMode.TailTruncation
     };
-    
+
+    private ActivityIndicator m_activityIndicator;
+
     public MultiLineInputField()
     {
         CreateButtons();
         SetupLabel();
+        CreateSaveView();
+    }
+
+    private void CreateSaveView()
+    {
+        m_activityIndicator = new Components.Loading.ActivityIndicator
+        { 
+            IsRunning = true, 
+            Opacity = 0,
+        };
+        
+        ContentBorderGrid.Add(m_activityIndicator);
     }
 
     protected override void CreateInputView()
@@ -87,12 +104,7 @@ public partial class MultiLineInputField : SingleLineInputField
         
         m_label.IsVisible = false;
 
-        if (InnerGrid.Contains(m_buttonsLayout))
-        {
-            InnerGrid.Remove(m_buttonsLayout);
-        }
-        
-        InnerGrid.Add(m_buttonsLayout, 0, 2);
+        ToggleButtonsVisibility(true);
         
         if(m_textWhenFirstFocused is not null)
             return;
@@ -118,10 +130,15 @@ public partial class MultiLineInputField : SingleLineInputField
         Reset();
     }
 
-    private void RemoveButtons()
+    private void ToggleButtonsVisibility(bool isEnabled)
     {
-        if(InnerGrid.Contains(m_buttonsLayout))
-            InnerGrid.Remove(m_buttonsLayout);
+        if (!InnerGrid.Contains(m_buttonsLayout))
+        {
+            InnerGrid.Add(m_buttonsLayout, 0, 2);
+            return;
+        }
+        
+        m_buttonsLayout!.IsVisible = isEnabled;
     }
 
     private void OnButtonTapped(bool isSaving)
@@ -129,7 +146,7 @@ public partial class MultiLineInputField : SingleLineInputField
         if (isSaving)
         {
             m_textWhenFirstFocused = InputView?.Text;
-            SaveCommand?.Execute(null);
+            SaveCommand?.Execute(SaveCommandParameter);
         }
         else
         {
@@ -164,7 +181,117 @@ public partial class MultiLineInputField : SingleLineInputField
 
     private void Reset()
     {
-        RemoveButtons();
+        ToggleButtonsVisibility(false);
         m_textWhenFirstFocused = null;
+    }
+
+    private void OnIsSavingChanged()
+    {
+        if (!IsSaving)
+            return;
+
+        OnIsSaving();
+    }
+
+    private async void OnIsSaving()
+    {
+        FadeOutContent();
+            
+        m_activityIndicator.IsVisible = true;
+        _ = m_activityIndicator.FadeTo(1);
+
+        // A small delay because OnUnfocused sets touch enabled to true
+        await Task.Delay(1);
+        SetTouchEnabled(false);
+    }
+    
+    private void OnIsSavingSuccessChanged()
+    {
+        if (!IsSavingSuccess)
+            return;
+
+        OnStopSaving(true);
+    }
+
+    private async void OnStopSaving(bool success)
+    {
+        FadeInContent();
+
+        SetTouchEnabled(true);
+
+        if (success)
+        {
+            Style = Styles.GetInputFieldStyle(InputFieldStyle.Success);
+        }
+        
+        await m_activityIndicator.FadeTo(0);
+        m_activityIndicator.IsVisible = false;
+    }
+
+    private void OnIsErrorChanged()
+    {
+        SetStyle();
+        ChangeHeaderTextStyle();
+        
+        if (IsError)
+        {
+            OnError();
+        }
+        else
+        {
+            OnChangedToNoError();
+        }
+    }
+
+    private void OnError()
+    {
+        OnStopSaving(false);
+        HelpTextLabel.SetBinding(Label.TextProperty, new Binding(nameof(ErrorText), source: this));
+        HelpTextLabel.TextColor = Colors.GetColor(ColorName.color_error_dark);
+    }
+
+    private void OnChangedToNoError()
+    {
+        HelpTextLabel.SetBinding(Label.TextProperty, new Binding(nameof(HelpText), source: this));
+        HelpTextLabel.SetBinding(Label.TextColorProperty, new Binding(nameof(HelpTextColor), source: this));
+        OnPropertyChanged(nameof(HelpText));
+    }
+
+    protected override void ChangeHeaderTextStyle()
+    {
+        base.ChangeHeaderTextStyle();
+
+        if (IsError)
+        {
+            if(IsFocused)
+                HeaderTextLabel.TextColor = Colors.GetColor(ColorName.color_error_dark);
+            
+            if(Text != string.Empty)
+                HeaderTextLabel.TextColor = Colors.GetColor(ColorName.color_error_dark);
+        }
+        else
+        {
+            HeaderTextLabel.TextColor = Colors.GetColor(ColorName.color_neutral_70);
+        }
+    }
+
+    protected override void SetStyle()
+    {
+        base.SetStyle();
+
+        if (IsError && !IsFocused)
+        {
+            Style = Styles.GetInputFieldStyle(InputFieldStyle.Error);
+        }
+    }
+
+    private void FadeOutContent()
+    {
+        InnerGrid.FadeTo(.25);
+    }
+
+    private void FadeInContent()
+    {
+        InnerGrid.FadeTo(1);
     }
 }