+++ src/library/DIPS.Mobile.UI/Resources/Styles/InputField/InputFieldStyle.cs
@@ -5,5 +5,7 @@ public enum InputFieldStyle
     None=0,
     Default=1,
     Focused=2,
-    Disabled=3
+    Disabled=3,
+    Error=4,
+    Success=5
 }